refactor: Moving the response file and its folder to the application layer
